% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blasso}
\alias{blasso}
\title{stagewise lasso}
\usage{
blasso(X, y, eps, tol, max_iteration)
}
\arguments{
\item{X}{predictor matrix}

\item{y}{response vector}

\item{eps}{stepsize}

\item{tol}{tolanrance paramater}

\item{max_iteration}{maximum iteration number}
}
\value{
A list consists of 
\item{beta_path}{the solution path of beta}
\item{lambda_path}{the path of lambda} 
\item{sigma2_path}{the path of estimated variance via maximum likelihood estimation}
}
\description{
stagewise lasso
}
