% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{codelasso}
\alias{codelasso}
\title{coordinate descent algorithm for lasso}
\usage{
codelasso(X, y, hat_beta, lambda, max_iteration, tol)
}
\arguments{
\item{X}{predictor matrix}

\item{y}{response vector}

\item{hat_beta}{a warm start for iteration}

\item{lambda}{penalty parameter}

\item{max_iteration}{maximum iteration number}

\item{tol}{tolanrance parameter}
}
\value{
hat_beta lasso estimation with lambda
}
\description{
coordinate descent algorithm for lasso
}
