% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{codelasso_path}
\alias{codelasso_path}
\title{lasso solution path via coordinate descent algorithm}
\usage{
codelasso_path(X, y, nlambda, max_iteration, tol, min_lambda)
}
\arguments{
\item{X}{predictor matrix}

\item{y}{response vector}

\item{nlambda}{the number of penalty parameter}

\item{max_iteration}{maximum iteration number}

\item{tol}{tolanrance parameter}

\item{min_lambda}{minimum lambda}
}
\value{
A list consists of
\item{beta_path}{the solution path of beta}
\item{lambda_path}{the lambda path}
\item{sigma2_path}{the path of estimated variance via maximum likelihood estimation}
}
\description{
lasso solution path via coordinate descent algorithm
}
