% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassopath.R
\name{path.control}
\alias{path.control}
\title{configure control list for lasso path}
\usage{
path.control(
  algo = "blasso",
  eps = 0.01,
  tol = 1e-06,
  max_iteration = 5000,
  nlambda = 100,
  min_lambda = 0
)
}
\arguments{
\item{algo}{algorithm type where "blasso" and "code" mean stagewise and coordinate descent algorithm, respectively.}

\item{eps}{stepsize for stagewise algorithm}

\item{tol}{tolanrance parameter}

\item{max_iteration}{maximum iteration number}

\item{nlambda}{the number of penalty parameter for coordinate descent algorithm}

\item{min_lambda}{minimum penalty parameter for coordinate descent algorithm}
}
\value{
A list consists of 
\item{algo}{algorithm type}
\item{eps}{stepsize for stagewise algorithm}
\item{tol}{tolanrance parameter} 
\item{max_iteration}{maximum iteration number}
\item{nlambda}{the number of penalty parameter for coordinate descent algorithm}
\item{min_lambda}{minimum penalty parameter for coordinate descent algorithm}
}
\description{
configure control list for lasso path
}
